Map editor now zooms exponentially instead of linearly
Zooming in UIs is typically implemented like this. In order for the
perceived zoom "speed" to be the same at every zoom level, the zoom
delta needs to be multiplied with the current zoom level, not added to.
Also, we apply an exponential function to the delta so that the zoom
speed is proportional to the speed at which the mouse wheel or touchpad
is moved.