Fix crash in web builds when undocking a map editor tab
This commit adds guards to the map, events and tilepicker so that they
don't render when even a small part of them is offscreen, because this
is not allowed in WebGPU or in future native builds of wgpu.

They aren't normally supposed to go offscreen, but it seems undocking
may cause temporary layout glitches for a frame.