Implement more graceful handling of offscreen viewports
It turns out it's possible for the map to be offscreen if the user
undocks the map editor and then shrinks the screen to be smaller than
the map editor.

Instead of simply refusing to render when this happens, we can just clip
the viewport to the screen and then offset the projection matrix.