Implement project drag and drop for Firefox
I implemented some code for cloning projects that are dragged and
dropped onto the program in Firefox into the Origin Private File System.
Since Firefox doesn't let you write to local files this is required.

My idea is we can clone the project into OPFS so that Luminol can edit
the files inside of OPFS and then the user can export the entire edited
project as a ZIP file later.

Unfortunately this doesn't seem to be practical because OPFS seems to be
too slow to handle cloning projects more than a few megabytes large.

The other option would be to lazily clone files as they're written to,
but this makes things hugely inconvenient if the files also have to be
modified elsewhere.

That's why I'm shelving this branch for now.