Fix `egui_dock::DockArea` causing parent to enlarge on each frame
This commit fixes a bug where dragging around tabs causes the parent
widget to continuously increase in height if the parent widget does not
have a fixed height.

This bug was manifesting in the script editor and the common events
editor.

egui_dock already has an issue for this at
https://github.com/Adanos020/egui_dock/issues/90